add Ethereum & Ethereum classic address generator
