+++ WalletKit/AddressFactory.swift
@@ -10,12 +10,18 @@ import Foundation
 
 class AddressFactory {
     static func provider(network: Network, publicKey: PublicKey) -> AddressProvider {
-        return BitcoinAddressProvider(network: network, publicKey: publicKey)
+        switch network.coin {
+        case .bitcoin:
+            return BitcoinAddressProvider(network: network, publicKey: publicKey)
+        case .ethereum, .ethereumClassic:
+            return EthereumAddressProvider(network: network, publicKey: publicKey)
+        }
+        
     }
 }
 
 public protocol AddressProvider {
     init(network: Network, publicKey: PublicKey)
-    var address: String { get }
+    var address: String? { get }
 }
